# Performance tweaks for List#diff, Seq#diff and Seq#intersect