@SethTisue 2021-01-15
@scala/collections second reviewer?