@SethTisue 2021-02-05
@scala/collections opinions here? (and/or volunteers to do some benchmarking?)