@martijnhoekstra 2021-05-25
@som-snytt I suspect the performance wins were lost in https://github.com/scala/scala/pull/9624 where much of this PR was lost to mima's relentlessness.