@martijnhoekstra 2021-05-25
Is your regression in terms of performance of correctness @som-snytt? Do you have some more details?