@som-snytt 2021-05-25
Yes in the linked ticket, `Array.empty[Double] intersect xs` throws because even an empty array knows its component. The OPT on isEmpty uses ArraySeq.empty.